% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.r
\name{CreateUKBiobankPhentoypes}
\alias{CreateUKBiobankPhentoypes}
\title{CreateUKBiobankPhentoypes - Function}
\usage{
CreateUKBiobankPhentoypes(Nvisits, visitreference, UKbioDataset,
  dfmaster_SQL_merge, dfDefinitions, Outputdir,
  VctOutputIndividualColumns = c())
}
\arguments{
\item{Nvisits}{#Number of follow-up visits in your dataset (a.t.m 3, baseline,1 and 2) .}

\item{visitreference}{#Number visit that should be used as reference; should be 0 (baseline),1 or 2}

\item{dfmaster_SQL_merge}{Dataframe: The Hesin table of UKBiobank that can be obtained by SQL; load via (WORK IN PROGRESS)}

\item{Outputdir}{Output directory}

\item{VctOutputIndividualColumns}{set VctOutputIndividualColumns=c("TS","SR","TS_RX","RX","LAB") to output all individual columns, default is 'VctOutputIndividualColumns=c()'}

\item{dfUKbioDataset}{Dataframe:The main UKBiobank dataset, currently only STATA files are allowed that are processed using the ukb_conv tool. Load with as.data.frame(read.dta13(file,convert.dates = TRUE,convert.factors=FALSE))}

\item{dfTraitTable}{Dataframe: A Trait-definition table. (EXAMPLE: WORK IN PROGRESS)}
}
\description{
CreateUKBiobankPhentoypes creates dichotomous phenotypes for UK Biobank using definitions that are based on the available variables (e.g. HES-data, mortality, toucschreen,  SR data). The current output includes variables on history, study visit, future, time-to-first-event; WORK IN PROGRESS
}
\examples{

library(gdata)
library(readstata13)
library(CreateUKBiobankPhentoypes)
Nvisits=3
visitreference=0

Outputdir<-getwd()
dfUKbioDataset<-as.data.frame(read.dta13("/path/to/UKbioDataset_file.dta",convert.dates = TRUE,convert.factors=FALSE))

hesin_file="/path/to/hesin-4july2016.tsv"
hesin_diagicd10_file="/path/to/hesin_diagicd10-4july2016.tsv"
hesin_diagicd9_file="/path/to/hesin_diagicd9_22_03_2016.tsv"
hesin_oper_file="/path/to/hesin_oper-4july2016.tsv"
dfhesintables<-LoadHesinTable(dfUKbioDataset,hesin_file,hesin_diagicd10_file,hesin_diagicd9_file,hesin_oper_file)

## load phenotype definitions.
data(dfDefinitions)




#run with minimum variables:
CreateUKBiobankPhentoypes(Nvisits,visitreference,UKbioDataset,dfhesintables,dfDefinitions,Outputdir)
#which is the same as:
CreateUKBiobankPhentoypes(Nvisits,visitreference,UKbioDataset,dfhesintables,dfDefinitions,Outputdir,VctOutputIndividualColumns=c(),VctAllUKBVDefinitionColumns=c("TS","SR","TS_RX","RX","LAB") )

#to output the variables based on RX (medication) in an *_RX file and exclude RX from the aggregate *_UKBV files run the following:
CreateUKBiobankPhentoypes(Nvisits,visitreference,UKbioDataset,dfhesintables,dfDefinitions,Outputdir,VctOutputIndividualColumns=c("RX",VctAllUKBVDefinitionColumns=c("TS","SR","TS_RX","LAB") )
}
\references{
<Please contact mail@niekverweij.com, this package is not yet ready to be distributed.>
}
\keyword{CreateUKBiobankPhentoypes}
