% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA-Codesheet.r
\docType{data}
\name{dfCodesheetREAD_SR.Coding}
\alias{dfCodesheetREAD_SR.Coding}
\title{Phenotype definitions}
\format{A dataframe}
\source{
niek verweij
}
\usage{
data(dfCodesheetREAD_SR.Coding)
}
\description{
Codesheet for translating between READ codes,  Self reported UKB codes and drug names during nurse interview.
}
\examples{
data(dfCodesheetREAD_SR.Coding)
}
\references{
tbd
}
\keyword{definitions}
