% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCodesheetTreatment-data.r
\docType{data}
\name{dfCodesheetTreatment}
\alias{dfCodesheetTreatment}
\title{Phenotype definitions}
\format{A dataframe}
\source{
niek verweij
}
\usage{
data(dfCodesheetTreatment)
}
\description{
Codesheet of treatments
}
\examples{
data(dfCodesheetTreatment)
}
\references{
tbd
}
\keyword{definitions}
