% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfDefinitions-data.r
\docType{data}
\name{dfDefinitions}
\alias{dfDefinitions}
\title{Phenotype definitions}
\format{A dataframe}
\source{
niek verweij
}
\usage{
data(dfDefinitions)
}
\description{
A set of phenotype defitions for UK Biobank
}
\examples{
data(dfDefinitions)
}
\references{
tbd
}
\keyword{definitions}
