% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.r
\name{ExtractVarsFromMasterSet}
\alias{ExtractVarsFromMasterSet}
\title{Extracting Vars From larger UK Bio dataset}
\usage{
ExtractVarsFromMasterSet(UKbioDataset, dfDefinitions_VarSelection,
  StataOutputFile)
}
\arguments{
\item{UKbioDataset}{df}

\item{dfDefinitions_VarSelection}{df}

\item{StataOutputFile}{file}
}
\value{
None
}
\description{
Extracts vars from ukbiobank from larger set of variables based on selection.
}
\examples{
ExtractVarsFromMasterSet(UKbioDataset,dfDefinitions_VarSelection,StataOutputFile)

}
\keyword{CreateUKBiobankPhentoypes}
\keyword{ExtractVarsFromMasterSet}
