% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-hesin.r
\name{LoadHesinTable}
\alias{LoadHesinTable}
\title{LoadHesinTable}
\usage{
LoadHesinTable(UKbioDataset, hesin_file, hesin_diagicd10_file,
  hesin_diagicd9_file, hesin_oper_file)
}
\arguments{
\item{hesin_file}{containing tsv file primary diagnoses for in hospital data, diagnoses + operation}

\item{hesin_diagicd10_file}{tsv file containing data on secondary ICD10 codes  for in hospital data}

\item{hesin_diagicd9_file}{tsv file containing data on secondary ICD9 codes for in hospital data}

\item{hesin_oper_file}{tsv file containing data on secondary operation codes for in hospital data}

\item{dfUKbioDataset}{Dataframe:The main UKBiobank dataset, currently only STATA files are allowed that are processed using the ukb_conv tool. Load with as.data.frame(read.dta13(file,convert.dates = TRUE))}
}
\description{
Combining Hesin tables for analyses by CreateUKBiobankPhentoypes()
}
\examples{


}
\keyword{LoadHesinTable}
