% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-hesin.r
\name{LoadHesinTable_v2}
\alias{LoadHesinTable_v2}
\title{LoadHesinTable_v2, v2 of uk biobank in hospital tables}
\usage{
LoadHesinTable_v2(UKbioDataset, fhesin, fhesin_diag, fhesin_oper)
}
\arguments{
\item{fhesin}{containing tsv file hes records}

\item{fhesin_diag}{tsv file containing data on secondary ICD10/ICD9 codes}

\item{fhesin_oper}{tsv file containing data on secondary OPCS codes}

\item{dfUKbioDataset}{Dataframe:The main UKBiobank dataset, currently only STATA files are allowed that are processed using the ukb_conv tool. Load with as.data.frame(read.dta13(file,convert.dates = TRUE))}
}
\description{
Combining Hesin tables for analyses by CreateUKBiobankPhentoypes()
}
\examples{


}
\keyword{LoadHesinTable_v2}
